# Serial flush faster