--- a/arduino/opencm_arduino/opencm9.04/variants/OpenCM904/hw/usb_cdc/usbd_cdc_interface.h
+++ b/arduino/opencm_arduino/opencm9.04/variants/OpenCM904/hw/usb_cdc/usbd_cdc_interface.h
@@ -74,7 +74,8 @@ uint8_t  CDC_Itf_Getch( void );
 uint8_t  CDC_Itf_Read( void );
 bool     CDC_Itf_IsConnected(void);
 int32_t  CDC_Itf_Peek( void );
-BOOL     CDC_Itf_IsTxTransmitted( void );
+void     CDC_Itf_flush_tx ( void );
+
 
 /* Exported macro ------------------------------------------------------------*/
 /* Exported functions ------------------------------------------------------- */