@KurtE 2018-08-23
Note: I have not included any update library file, not sure if this should just be a copy a generated file that is out of my temporary files?  

As asked in the thread: https://community.robotsource.org/t/opencm-9-04-how-to-build-using-the-develop-branch-and-make-pull-requests/1474