@OpusK 2018-08-26
Hi, @KurtE 

After modifying the code in [OpenCM9.04/arduino/opencm_arduino/opencm9.04/](https://github.com/ROBOTIS-GIT/OpenCM9.04/tree/master/arduino/opencm_arduino/opencm9.04) directory, build through EclipseIDE, a library file (.a) is created.

Then copy this file to [variants/OpenCM904](OpenCM9.04/arduino/opencm_arduino/opencm9.04_release/variants/OpenCM904/) in the release folder.
Also, if any .h files have been changed, you must also overwrite them in the corresponding header file location in the release folder.