@KurtE 2018-08-26
@OpusK,
I now have copied over the changed header files to the release folder. 

Not sure what is needed to setup to run this with Eclipse?  Are you using Sloeber?  Do you have some specific setup for it?  I tried installing latest Sloeber and tried adding to use my hardware folder, but it fails to find compiler... 

Does this generate a different library file than the Arduino IDE creates.  
That is if I run a build using Arduino on a simple sketch, it generates a library file (core.a) in this case in the folder:  C:\Users\kurte\AppData\Local\Temp\arduino_build_334393\core
Which in this latest run was about 324KB in size.   Which I am assuming is not the same thing as the
lib_f103.a that is currently on my system in the release folder is something like 68394KB which is over 200 times larger in size?