@OpusK 2018-08-27
@KurtE, 

The project that creates the library (.a) does not need the Arduino plugin.
It is simply generic firmware.

Currently, you need to use the Eclipse IDE. If you have a separate Makefile, you can do it without an IDE.

If you need more information about installing and configuring the Eclipse CDT IDE, please see the link below. I hope it helps.
 - [[Eclipse] GCC ARM develop environment + OpenOCD for Linux](https://blog.naver.com/opusk/221008480771)
 - [[Eclipse] GCC ARM develop environment + OpenOCD for Windows](https://blog.naver.com/opusk/221006348299)