@KurtE 2018-08-27
@OpusK,

Ok I believe I have finally built the archive file ;) 

Hopefully everything is valid.